restructure, add Clock and Car eviction

+++ variants/src/test/java/org/cache2k/benchmark/OptBenchmark.java
@@ -57,7 +57,7 @@ public class OptBenchmark extends RandomCacheBenchmark {
 
     @Override
     public void record(final Integer v) {
-      trace.add(v);
+      trace.add((int) v);
     }
 
     @Override
